Plot Mun manuever at the next available position in orbit
Previously we only set manuever nodes at the ascending or descending
node. Now we plot a manuever node 60 seconds into the future - if this
doesn't encounter the Mun's sphere of influence, plot another manuever
node further into the future. Repeat this process until we find a
manuever node that would encounter the Mun.